+++ src/main/java/org/accounting/system/validators/AccessProjectValidator.java
@@ -5,10 +5,12 @@ import io.vavr.control.Try;
 import jakarta.enterprise.inject.spi.CDI;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
+import org.accounting.system.beans.RequestUserContext;
 import org.accounting.system.constraints.AccessProject;
 import org.accounting.system.enums.Collection;
 import org.accounting.system.enums.Operation;
 import org.accounting.system.exceptions.CustomValidationException;
+import org.accounting.system.repositories.client.ClientRepository;
 import org.accounting.system.repositories.project.ProjectRepository;
 
 /**
@@ -43,6 +45,16 @@ public class AccessProjectValidator implements ConstraintValidator<AccessProject
             Try
                     .run(()->projectRepository.findByIdOptional(project).orElseThrow(()->new CustomValidationException("There is no Project with the following id: "+project, HttpResponseStatus.NOT_FOUND)))
                     .getOrElseThrow(()->new CustomValidationException("There is no Project with the following id: "+project, HttpResponseStatus.NOT_FOUND));
+        }
+
+        var clientRepository = CDI.current().select(ClientRepository.class).get();
+
+        var requestUserContext = CDI.current().select(RequestUserContext.class).get();
+
+
+        if(clientRepository.isSystemAdmin(requestUserContext.getId())){
+
+            return true;
         }
 
         var access =  projectRepository.accessibility(project, collection, operation);