+++ CHANGELOG.md
@@ -26,6 +26,12 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the
 
 ---
 
+## Unreleased - Added
+
+- [#219](https://github.com/ARGOeu/argo-accounting/pull/219) - ACC-345 Sys Admin 
+  should be able to manage all projects.
+
+
 ## 1.6.1 - 2025-03-21
 
 ---